fix(*): fix execute module factory compability
